requirements.txt
boto3==1.34.162
pydantic==2.8.2
PyYAML==6.0.1
tenacity
rich==13.7.1
python-dotenv==1.0.1
openai==1.42.0
httpx==0.27.2
fastapi==0.115.0
uvicorn==0.30.3
streamlit==1.37.0
requests==2.32.3
python-multipart==0.0.9
librosa==0.10.2.post1
soundfile==0.12.1
transformers==4.44.2